refactor: break out ViewControlMenu into its own component
